WIP: Conver the upload command to just shellout to vagrant
There are multiple bugs in the vagrant API gem that make it a less than desirable way to get the files up to Vagrant Cloud. I've fixed a few of them and reported others, but we need to get something working here so this just rips that out and uses the command instead. It's slower and more error prone, but it gets us something that works. If that gem gets improvements sometime down the line we can revisit this.

Signed-off-by: Tim Smith <tsmith@chef.io>